move insert_dataframe tests into the tests/numpy/ directory
